communication_service-dev/requirements.txt
--extra-index-url https://__token__:8zKkkbRtGyJ1LdxxUshA@gitlab.renewbuy.in/api/v4/projects/103/packages/pypi/simple
rb-utils==2023.2.17.7.49.53
aiofiles==0.8.0
aiohttp==3.8.1
aiokafka==0.7.2
aioredis==1.3.1
aiosmtplib==1.1.7
aiosqlite==0.17.0
alembic==1.8.0
anyio==3.6.1
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
asyncpg==0.25.0
backcall==0.2.0
black==22.6.0
blinker==1.5
certifi==2022.6.15
click==8.1.3
colorama==0.4.5
databases==0.6.0
decorator==5.1.1
dnspython==2.2.1
docstring-parser==0.12
email-validator==1.2.1
emails==0.6
executing==1.0.0
fastapi==0.78.0
fastapi-mail==1.1.4
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.2
h11==0.12.0
h2==4.1.0
hiredis==2.0.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
inflection==0.5.1
ipdb==0.13.9
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
kafka-python==2.0.2
Mako==1.2.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
ormar==0.11.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
piccolo==0.80.1
piccolo-admin==0.21.0
piccolo-api==0.38.0
pickleshare==0.7.5
platformdirs==2.5.2
priority==2.0.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.9.1
Pygments==2.13.0
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqladmin==0.1.11
SQLAlchemy==1.4.37
SqlAlchemy-Enum-Tables==1.1.0
SQLAlchemy-Enum34==2.0.0
SQLAlchemy-Utils==0.38.3
stack-data==0.5.0
starlette==0.19.1
targ==0.3.7
toml==0.10.2
tomli==2.0.1
traitlets==5.3.0
typing-extensions==4.2.0
uvicorn==0.17.6
wcwidth==0.2.5
wsproto==1.1.0
WTForms==3.0.1
zipp==3.8.0
